Split builders into separate files. Add template builder. Add ability to add an existing children builders to a group builder.
