+++ libs/formly-builder/src/index.ts
@@ -1,3 +1,3 @@
 export * from './lib/formly-builder.module';
-export * from './lib/formly-builder';
+export * from './lib/builders';
 export * from './lib/operators';
