+++ libs/formly-builder/src/lib/builders/builder.ts
@@ -1,0 +1,3 @@
+export interface Builder<T> {
+  build(): T;
+}
