+++ libs/formly-builder/src/lib/builders/group-builder.ts
@@ -1,37 +1,12 @@
 import { FormlyFieldConfig } from '@ngx-formly/core';
-import { pipeFromArray } from './pipe';
-import { MonoTypeOperatorFunction } from './types';
+import { pipeFromArray } from '../pipe';
+import { MonoTypeOperatorFunction, Obj } from '../types';
+import { Builder } from './builder';
+import { FieldBuilder } from './field-builder';
+import { TemplateBuilder } from './template-builder';
 
-export type Obj = { [key: string]: any };
 
-export interface Builder<T> {
-  build(): T;
-}
-
-export class FieldBuilder<T> implements Builder<FormlyFieldConfig> {
-
-  private operations: MonoTypeOperatorFunction<FormlyFieldConfig>[];
-
-  constructor(public key: T) { }
-
-  public build(): FormlyFieldConfig {
-    let props = {};
-    if (this.operations?.length > 0) {
-      props = pipeFromArray<FormlyFieldConfig>(this.operations)(props);
-    }
-
-    let result = {
-      ...props,
-      key: this.key as unknown as string
-    };
-    return result;
-  }
-
-  withProps(...operations: MonoTypeOperatorFunction<FormlyFieldConfig>[]): FieldBuilder<T> {
-    this.operations = operations;
-    return this;
-  }
-}
+export type ProjectorFn<R extends Builder<any>, K extends GroupBuildBase<any>> = (value: K) => R[];
 
 
 export abstract class GroupBuildBase<T extends Obj> {
@@ -39,25 +14,29 @@ export abstract class GroupBuildBase<T extends Obj> {
   protected _builders: Array<any> = [];
 
   public field<K extends keyof T>(key: K): FieldBuilder<any> {
-    return this.addBuilder(new FieldBuilder(key));
+    return this.add(new FieldBuilder(key));
   }
 
-  public withFields<R extends Builder<any>, K = this>(project: (value: K) => R[]): GroupBuildBase<T> {
-    this.addBuilder(project);
+  public withFields<R extends Builder<any>, K extends GroupBuildBase<T>>(project: ProjectorFn<R, K>): GroupBuildBase<T> {
+    this.add(project);
     return this;
   }
 
   public group<K extends keyof T>(key: K): GroupBuilder<T[K]> {
-    return this.addBuilder(new GroupBuilder<T[K]>(key));
+    return this.add(new GroupBuilder<T[K]>(key));
   }
 
   public layout(): LayoutBuilder<T> {
-    return this.addBuilder(new LayoutBuilder<T>());
+    return this.add(new LayoutBuilder<T>());
+  }
+
+  public template(template: string): TemplateBuilder {
+    return this.add(new TemplateBuilder(template));
   }
 
   abstract build(): any;
 
-  protected addBuilder<U>(value: U): U {
+  public add<U extends Builder<FormlyFieldConfig | FormlyFieldConfig[]> | ProjectorFn<Builder<any>, GroupBuildBase<any>>>(value: U): U {
     this._builders.push(value);
     return value;
   }
@@ -70,13 +49,13 @@ export abstract class GroupBuildBase<T extends Obj> {
         const projectorBuilders = builder(current) as Array<Builder<any>>;
         result.push(...projectorBuilders.map(f => f.build()));
       } else {
-        result.push(builder.build());
+        const config = builder.build();
+        Array.isArray(config) ? result.push(...config) : result.push(config);
       }
     });
 
     return result;
   }
-
 }
 
 export class GroupBuilder<T extends Obj> extends GroupBuildBase<T> implements Builder<FormlyFieldConfig> {
@@ -109,7 +88,6 @@ export class GroupBuilder<T extends Obj> extends GroupBuildBase<T> implements Bu
   }
 }
 
-
 export class LayoutBuilder<T> extends GroupBuildBase<T> {
 
   private operations: MonoTypeOperatorFunction<FormlyFieldConfig>[];
@@ -133,14 +111,4 @@ export class LayoutBuilder<T> extends GroupBuildBase<T> {
     this.operations = operations;
     return this;
   }
-
-}
-
-
-export class FormlyBuilder<T extends Obj = any> extends GroupBuildBase<T> implements Builder<FormlyFieldConfig[]> {
-
-  public build(): FormlyFieldConfig[] {
-    return this.buildGroup(this, this._builders);
-  }
-
 }
