+++ libs/formly-builder/src/lib/builders/index.ts
@@ -1,0 +1,4 @@
+export * from './formly-builder';
+export * from './field-builder';
+export * from './group-builder';
+export * from './template-builder';
